deprecate more functions
