+++ includes/functions-settings.php
@@ -22,11 +22,24 @@ add_action( 'init', 'ucare\register_settings' );
  */
 function register_settings() {
 
+    /**
+     * @since 1.0.0
+     */
     register_setting( 'uc-advanced', Options::TEMPLATE_PAGE_ID, array(
         'type' => 'integer',
         'sanitize_callback' => 'ucare\sanitize_post_id',
     ) );
 
+    register_setting( 'uc-general', Options::ECOMMERCE, array(
+        'type'    => 'string',
+        'default' => Defaults::ECOMMERCE,
+        'sanitize_callback' => 'ucare\sanitize_boolean'
+    ) );
+
+
+    /**
+     * @since 1.6.0
+     */
     register_setting( 'uc-advanced', Options::CREATE_TICKET_PAGE_ID, array(
         'type' => 'integer',
         'sanitize_callback' => 'ucare\sanitize_post_id',
@@ -40,12 +53,6 @@ function register_settings() {
     register_setting( 'uc-advanced', Options::LOGIN_PAGE_ID, array(
         'type' => 'integer',
         'sanitize_callback' => 'ucare\sanitize_post_id',
-    ) );
-
-    register_setting( 'uc-general', Options::ECOMMERCE, array(
-        'type'    => 'string',
-        'default' => Defaults::ECOMMERCE,
-        'sanitize_callback' => 'ucare\sanitize_boolean'
     ) );
 
 }
