+++ includes/functions.php
@@ -9,6 +9,32 @@ namespace ucare;
 
 
 /**
+ * Convert a hexadecimal value to RGB.
+ *
+ * @param string $hex
+ *
+ * @since 1.6.0
+ * @return array
+ */
+function hex2rgb( $hex ) {
+    $hex = str_replace( "#", "", $hex );
+
+    if ( strlen( $hex ) == 3 ) {
+        $r = hexdec( substr( $hex, 0, 1 ) . substr( $hex, 0, 1 ) );
+        $g = hexdec( substr( $hex, 1, 1 ) . substr( $hex, 1, 1 ) );
+        $b = hexdec( substr( $hex, 2, 1 ) . substr( $hex, 2, 1 ) );
+    } else {
+        $r = hexdec( substr( $hex, 0, 2 ) );
+        $g = hexdec( substr( $hex, 2, 2 ) );
+        $b = hexdec( substr( $hex, 4, 2 ) );
+    }
+    $rgb = array ( $r, $g, $b );
+    //return implode(",", $rgb); // returns the rgb values separated by commas
+    return $rgb; // returns an array with the rgb values
+}
+
+
+/**
  * Safely pluck a value from an object or array.
  *
  * @param object|array $obj
@@ -41,6 +67,54 @@ function pluck( $obj, $field, $default = false ) {
 }
 
 
+/**
+ * Check if a $_REQUEST nonce is valid.
+ *
+ * @param        $action
+ * @param string $nonce
+ *
+ * @since 1.4.2
+ * @return bool|false|int
+ */
+function verify_request_nonce( $action, $nonce = '_wpnonce' ) {
+
+    if ( isset( $_REQUEST[ $nonce ] ) ) {
+        return wp_verify_nonce( $_REQUEST[ $nonce ], $action );
+    }
+
+    return false;
+
+}
+
+
+/**
+ * Get a variable from the $_REQUEST.
+ *
+ * @param               $var
+ * @param string        $default
+ * @param callable|null $sanitize
+ *
+ * @since 1.4.2
+ * @return string
+ */
+function get_var( $var, $default = '', callable $sanitize = null ) {
+
+    if ( isset( $_REQUEST[ $var ] ) ) {
+        return !empty( $sanitize ) ? $sanitize( $_REQUEST[ $var ] ) : $_REQUEST[ $var ];
+    }
+
+    return $default;
+}
+
+
+/**
+ * Get a list of the MIME types that are allowed to be uploaded as attachments.
+ *
+ * @param null|string $type
+ *
+ * @since 1.4.2
+ * @return array
+ */
 function allowed_mime_types( $type = null ) {
 
     $file_types  = explode( ',', get_option( Options::FILE_MIME_TYPES, Defaults::FILE_MIME_TYPES   ) );
@@ -57,34 +131,11 @@ function allowed_mime_types( $type = null ) {
 }
 
 
-
-
-function selectbox( $name, $options, $selected = '', $attrs = array() ) { ?>
-
-    <select name="<?php esc_attr_e( $name ); ?>"
-
-            <?php foreach ( $attrs as $attr => $values ) : ?>
-
-                <?php echo $attr . '="' . esc_attr( $values ) . '"' ?>
-
-            <?php endforeach; ?>>
-
-        <?php foreach ( $options as $value => $label ) : ?>
-
-            <option value="<?php esc_attr_e( $value ); ?>"
-
-                <?php selected( $selected, $value ); ?> ><?php echo $label ?></option>
-
-        <?php endforeach; ?>
-
-    </select>
-
-<?php }
-
-
-
-
-
+/***********************************************************************************************************************
+ *
+ * Everything below this comment will be deprecated. Sub namespaces of ucare are no longer supported within core.
+ *
+ */
 
 
 namespace ucare\util;
@@ -93,36 +144,6 @@ use ucare\Options;
 use ucare\Plugin;
 
 
-function can_use_support( $id = false ) {
-    if( $id ) {
-
-        $result = user_can( $id, 'use_support' );
-    } else {
-        $result = current_user_can( 'use_support' );
-    }
-
-    return $result;
-}
-
-function can_manage_tickets( $id = false ) {
-    if( $id ) {
-        $result = user_can( $id, 'manage_support_tickets' );
-    } else {
-        $result = current_user_can( 'manage_support_tickets' );
-    }
-
-    return $result;
-}
-
-function can_manage_support( $id = false ) {
-    if( $id ) {
-        $result = user_can( $id, 'manage_support' );
-    } else {
-        $result = current_user_can( 'manage_support' );
-    }
-
-    return $result;
-}
 
 
 function extract_tags( $str, $open, $close ) {
@@ -191,56 +212,6 @@ function list_agents() {
 }
 
 
-function roles() {
-    return array(
-        'support_admin' => __( 'Support Admin', 'ucare' ),
-        'support_agent' => __( 'Support Agent', 'ucare' ),
-        'support_user'  => __( 'Support User', 'ucare' ),
-    );
-}
-
-function add_caps( $role, $privilege = '' ) {
-    $role = get_role( $role );
-
-    if( !empty( $role ) ) {
-        switch( $privilege ) {
-            case 'manage':
-                $role->add_cap( 'create_support_tickets' );
-                $role->add_cap( 'use_support' );
-                $role->add_cap( 'manage_support_tickets' );
-                $role->add_cap( 'edit_support_ticket_comments' );
-
-                break;
-
-            case 'admin':
-                $role->add_cap( 'create_support_tickets' );
-                $role->add_cap( 'use_support' );
-                $role->add_cap( 'manage_support_tickets' );
-                $role->add_cap( 'edit_support_ticket_comments' );
-                $role->add_cap( 'manage_support' );
-
-                break;
-
-            default:
-                $role->add_cap( 'create_support_tickets' );
-                $role->add_cap( 'use_support' );
-
-                break;
-        }
-    }
-}
-
-function remove_caps( $role ) {
-    $role = get_role( $role );
-
-    if( !empty( $role ) ) {
-        $role->remove_cap( 'create_support_tickets' );
-        $role->remove_cap( 'use_support' );
-        $role->remove_cap( 'manage_support_tickets' );
-        $role->remove_cap( 'edit_support_ticket_comments' );
-        $role->remove_cap( 'manage_support' );
-    }
-}
 
 
 
@@ -353,75 +324,9 @@ function create_email_templates() {
     }
 }
 
-function configure_roles() {
-    $administrator = get_role( 'administrator' );
 
-    $administrator->add_cap( 'read_support_ticket' );
-    $administrator->add_cap( 'read_support_tickets' );
-    $administrator->add_cap( 'edit_support_ticket' );
-    $administrator->add_cap( 'edit_support_tickets' );
-    $administrator->add_cap( 'edit_others_support_tickets' );
-    $administrator->add_cap( 'edit_published_support_tickets' );
-    $administrator->add_cap( 'publish_support_tickets' );
-    $administrator->add_cap( 'delete_support_tickets' );
-    $administrator->add_cap( 'delete_others_support_tickets' );
-    $administrator->add_cap( 'delete_private_support_tickets' );
-    $administrator->add_cap( 'delete_published_support_tickets' );
 
-    foreach( \ucare\util\roles() as $role => $name ) {
-        add_role( $role, $name );
-    }
 
-    \ucare\util\add_caps( 'customer' );
-    \ucare\util\add_caps( 'subscriber' );
-    \ucare\util\add_caps( 'support_user' );
-
-    \ucare\util\add_caps( 'support_agent' , 'manage' );
-
-    \ucare\util\add_caps( 'support_admin' , 'admin' );
-    \ucare\util\add_caps( 'administrator' , 'admin' );
-}
-
-function cleanup_roles() {
-    foreach( \ucare\util\roles() as $role => $name ) {
-        remove_role( $role );
-    }
-
-    \ucare\util\remove_caps( 'customer' );
-    \ucare\util\remove_caps( 'subscriber' );
-    \ucare\util\remove_caps( 'administrator' );
-
-    $administrator = get_role( 'administrator' );
-
-    $administrator->remove_cap( 'read_support_ticket' );
-    $administrator->remove_cap( 'read_support_tickets' );
-    $administrator->remove_cap( 'edit_support_ticket' );
-    $administrator->remove_cap( 'edit_support_tickets' );
-    $administrator->remove_cap( 'edit_others_support_tickets' );
-    $administrator->remove_cap( 'edit_published_support_tickets' );
-    $administrator->remove_cap( 'publish_support_tickets' );
-    $administrator->remove_cap( 'delete_support_tickets' );
-    $administrator->remove_cap( 'delete_others_support_tickets' );
-    $administrator->remove_cap( 'delete_private_support_tickets' );
-    $administrator->remove_cap( 'delete_published_support_tickets' );
-}
-
-function hex2rgb( $hex ) {
-    $hex = str_replace( "#", "", $hex );
-
-    if ( strlen( $hex ) == 3 ) {
-        $r = hexdec( substr( $hex, 0, 1 ) . substr( $hex, 0, 1 ) );
-        $g = hexdec( substr( $hex, 1, 1 ) . substr( $hex, 1, 1 ) );
-        $b = hexdec( substr( $hex, 2, 1 ) . substr( $hex, 2, 1 ) );
-    } else {
-        $r = hexdec( substr( $hex, 0, 2 ) );
-        $g = hexdec( substr( $hex, 2, 2 ) );
-        $b = hexdec( substr( $hex, 4, 2 ) );
-    }
-    $rgb = array ( $r, $g, $b );
-    //return implode(",", $rgb); // returns the rgb values separated by commas
-    return $rgb; // returns an array with the rgb values
-}
